Move _DimAllocator into indep_poutine.py
